Specify required Python version via caret notation
Specifying a minimum version instead would normally be preferred, but
that resulted in dependency resolution errors when other dependent
projects ostensibly used caret notation (^3.6) that translate into
implicit bracketed requirements (such as >=3.6, <4.0). Hard to tell
whether this is a Poetry bug -- since an error shouldn't result from not
specifying an upward bound -- but for now the easiest fix is to play
along and hope this is fixed so we can go back to just specifying a
minimum version.